+++ src/neptune_optuna/__init__.py
@@ -14,7 +14,9 @@
 # limitations under the License.
 #
 
-from ._version import get_versions
-
-__version__ = get_versions()["version"]
-del get_versions
+from neptune_optuna.impl import (
+    NeptuneCallback,
+    __version__,
+    load_study_from_run,
+    log_study_metadata,
+)